Move the bulk edit custom box code to the controller
